+++ __tests__/src/optics/cardinalElements.test.ts
@@ -3,11 +3,68 @@ import buildLens from "../../../src/optics/buildLens.js";
 import { computeCardinalElementsAtState } from "../../../src/optics/cardinalElements.js";
 import { doLayout } from "../../../src/optics/optics.js";
 import { LENS_CATALOG } from "../../../src/utils/catalog/lensCatalog.js";
-import type { RuntimeLens } from "../../../src/types/optics.js";
+import type { RuntimeLens, SurfaceData } from "../../../src/types/optics.js";
 
 function cardinalsFor(L: RuntimeLens, focusT = 0, zoomT = 0) {
   const layout = doLayout(focusT, zoomT, L);
   return computeCardinalElementsAtState(L, focusT, zoomT, layout.z, layout.imgZ);
+}
+
+function minimalCardinalLens(surfaces: SurfaceData[]): RuntimeLens {
+  const imageZ = surfaces.reduce((sum, surface) => sum + surface.d, 0);
+  return {
+    data: {
+      key: "cardinal-fixture",
+      name: "Cardinal fixture",
+      closeFocusM: 0.5,
+      focusStep: 0.004,
+      maxFstop: 16,
+      apertureStep: 0.004,
+      fstopSeries: [2, 2.8, 4, 5.6, 8, 11, 16],
+      elements: [],
+      surfaces,
+      svgW: 800,
+      svgH: 400,
+      scFill: 0.55,
+      yScFill: 0.55,
+      clipMargin: 1,
+      maxRimAngleDeg: 64,
+      gapSagFrac: 0.9,
+      maxAspectRatio: 1.6,
+      rayFractions: [-0.5, 0, 0.5],
+      rayLeadFrac: 0.35,
+      offAxisFieldFrac: 0.6,
+      offAxisFractions: [-0.5, 0, 0.5],
+      nominalFno: 2,
+    },
+    S: surfaces,
+    N: surfaces.length,
+    ES: [],
+    elements: [],
+    asphByIdx: {},
+    varByIdx: {},
+    vdByIdx: {},
+    spectralByIdx: {},
+    indexByIdx: {},
+    varLabels: [],
+    groups: [],
+    doublets: [],
+    perspectiveControl: null,
+    aberrationControl: null,
+    projection: { kind: "rectilinear" },
+    opticalPath: { mode: "sequential", surfaceOrder: null, surfaceLabels: null, maxInteractions: surfaces.length + 1 },
+    imagePlane: { z: imageZ, y: 0, normal: { z: 1, y: 0 }, label: "IMG" },
+    isFoldedOptics: false,
+    stopIdx: 0,
+    stopPhysSD: surfaces[0]?.sd ?? 1,
+    isZoom: false,
+    zoomPositions: null,
+    zoomLabels: null,
+    zoomStep: 0.01,
+    rayFractions: [-0.5, 0, 0.5],
+    offAxisFractions: [-0.5, 0, 0.5],
+    offAxisFieldFrac: 0.6,
+  } as unknown as RuntimeLens;
 }
 
 describe("computeCardinalElementsAtState", () => {
@@ -25,12 +82,7 @@ describe("computeCardinalElementsAtState", () => {
   });
 
   it("computes nodal points independently when image-side index differs", () => {
-    const L = {
-      S: [{ label: "1", R: 100, d: 0, nd: 1.5, elemId: 1, sd: 10 }],
-      varByIdx: {},
-      isZoom: false,
-      N: 1,
-    } as unknown as RuntimeLens;
+    const L = minimalCardinalLens([{ label: "1", R: 100, d: 0, nd: 1.5, elemId: 1, sd: 10 }]);
 
     const result = computeCardinalElementsAtState(L, 0, 0, [0], 20);
 
@@ -83,12 +135,7 @@ describe("computeCardinalElementsAtState", () => {
   });
 
   it("returns null for afocal or zero-power systems", () => {
-    const L = {
-      S: [{ label: "1", R: 1e15, d: 0, nd: 1, elemId: 1, sd: 10 }],
-      varByIdx: {},
-      isZoom: false,
-      N: 1,
-    } as unknown as RuntimeLens;
+    const L = minimalCardinalLens([{ label: "1", R: 1e15, d: 0, nd: 1, elemId: 1, sd: 10 }]);
 
     expect(computeCardinalElementsAtState(L, 0, 0, [0], 10)).toBeNull();
   });
