+++ src/components/hooks/useOnAxisRays.ts
@@ -7,7 +7,7 @@
  * show the extrapolated path of rays clipped by the aperture stop.
  */
 import { useMemo } from "react";
-import { getSelectedOpticsEngine } from "../../optics/engineSelector.js";
+import { traceRay } from "../../optics/optics.js";
 import { obstructionAwareRayFractionsForDensity } from "../../optics/raySampling.js";
 import type { RuntimeLens } from "../../types/optics.js";
 import type { RayDensity } from "../../types/state.js";
@@ -61,8 +61,6 @@ export default function useOnAxisRays({
   focusK,
   lensKey,
 }: UseOnAxisRaysParams): UseOnAxisRaysResult {
-  const opticsEngine = getSelectedOpticsEngine();
-
   return useMemo((): UseOnAxisRaysResult => {
     if (!L) return { segments: [], error: null };
     try {
@@ -70,7 +68,7 @@ export default function useOnAxisRays({
       for (const f of obstructionAwareRayFractionsForDensity(L, L.rayFractions, rayDensity, currentEPSD)) {
         const h = f * currentEPSD;
         const uIn = rayTracksF ? h * focusK : 0;
-        const rawResult = opticsEngine.traceRay(h, uIn, zPos, focusT, zoomT, currentPhysStopSD, true, L, aberrationT);
+        const rawResult = traceRay(h, uIn, zPos, focusT, zoomT, currentPhysStopSD, true, L, aberrationT);
         const result = movementTransform ? movementTransform.trace(rawResult) : rawResult;
         out.push(
           compileRaySegment(
@@ -109,6 +107,5 @@ export default function useOnAxisRays({
     lensKey,
     clampedRayEnd,
     movementTransform,
-    opticsEngine,
   ]);
 }
