+++ src/optics/chiefRayDiagnostics.ts
@@ -1,32 +1,6 @@
-import type { ChiefRaySolveResult } from "./fieldGeometry.js";
-
-export type ChiefRayStatus = ChiefRaySolveResult["status"];
-
-export type ChiefRayStatusCounts = Record<ChiefRayStatus, number>;
-
-const counts: Map<string, ChiefRayStatusCounts> = new Map();
-
-function emptyCounts(): ChiefRayStatusCounts {
-  return { converged: 0, "paraxial-fallback": 0, "bracket-failed": 0, "out-of-domain": 0 };
-}
-
-export function recordChiefRayStatus(lensKey: string, status: ChiefRayStatus): void {
-  let entry = counts.get(lensKey);
-  if (!entry) {
-    entry = emptyCounts();
-    counts.set(lensKey, entry);
-  }
-  entry[status] += 1;
-}
-
-export function getChiefRayDiagnostics(): Map<string, ChiefRayStatusCounts> {
-  const snapshot = new Map<string, ChiefRayStatusCounts>();
-  for (const [key, value] of counts.entries()) {
-    snapshot.set(key, { ...value });
-  }
-  return snapshot;
-}
-
-export function resetChiefRayDiagnostics(): void {
-  counts.clear();
-}
+export {
+  getChiefRayDiagnostics2 as getChiefRayDiagnostics,
+  resetChiefRayDiagnostics2 as resetChiefRayDiagnostics,
+  type ChiefRayDiagnostics2 as ChiefRayStatusCounts,
+  type ChiefRayStatus2 as ChiefRayStatus,
+} from "../optics-2/compat.js";
